[EmbeddingAPI] Move some cases to v6
-Move 6 cases to v6 & update related files
-Cover the XWalkActivity and XWalkInitializer two ways

Impacted tests(approved): new 0, update 6, delete 0
Unit test platform: [Android]
Unit test result summary: pass 6, fail 0, block 0
