+++ embeddingapi/embedding-api-android-tests/tests.full.xml
@@ -83,7 +83,7 @@
           <test_script_entry>org.xwalk.embedding.test.v4.XWalkResourceClientTest</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTest" platform="android" priority="P1" purpose="Check if the methods of XWalkView interface can be executed correctly." status="approved" type="functional_positive" subcase="16">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTest" platform="android" priority="P1" purpose="Check if the methods of XWalkView interface can be executed correctly." status="approved" type="functional_positive" subcase="12">
         <description>
           <test_script_entry>org.xwalk.embedding.test.v5.XWalkViewTest</test_script_entry>
         </description>
@@ -116,6 +116,11 @@
       <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkSettingTest" platform="android" priority="P1" purpose="Check if the methods of XWalkSetting interface can be executed correctly." status="approved" type="functional_positive" subcase="7">
         <description>
           <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTest</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkViewTest" platform="android" priority="P1" purpose="Check if the methods of XWalkView interface can be executed correctly." status="approved" type="functional_positive" subcase="3">
+        <description>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkViewTest</test_script_entry>
         </description>
       </testcase>
     </set>
