+++ embeddingapi/embedding-api-android-tests/tests_v5.xml
@@ -3,7 +3,7 @@
 <test_definition>
   <suite name="webapi-embeddingapi-xwalk-tests" category="Android embedding APIs">
     <set name="EmbeddingApiTest" type="androidunit" location="device">
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTest" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="16">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTest" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="12">
         <description>
           <test_script_entry>org.xwalk.embedding.test.v5.XWalkViewTest</test_script_entry>
         </description>
