+++ embeddingapi/embedding-api-android-tests/tests_v6.xml
@@ -3,6 +3,11 @@
 <test_definition>
   <suite name="webapi-embeddingapi-xwalk-tests" category="Android embedding APIs">
     <set name="EmbeddingApiTest" type="androidunit" location="device">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkViewTest" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="3">
+        <description>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkViewTest</test_script_entry>
+        </description>
+      </testcase>
       <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkSettingTest" purpose="Check if the methods of XWalkSetting interface can be executed correctly."  subcase="7">
         <description>
           <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTest</test_script_entry>
