+++ embeddingapi/embedding-asyncapi-android-tests/tests.full.xml
@@ -118,6 +118,11 @@
           <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTestAsync</test_script_entry>
         </description>
       </testcase>
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkViewTestAsync" platform="android" priority="P1" purpose="Check if the methods of XWalkView interface can be executed correctly." status="approved" type="functional_positive" subcase="3">
+        <description>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkViewTestAsync</test_script_entry>
+        </description>
+      </testcase>
     </set>
   </suite>
 </test_definition>
