+++ embeddingapi/embedding-asyncapi-android-tests/tests_v5.xml
@@ -3,7 +3,7 @@
 <test_definition>
   <suite name="Embedding-asyncapi-android-tests" category="Android embedding APIs">
     <set name="EmbeddingApiAsyncTest" type="androidunit" location="device">
-      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTestAsync" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="16">
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v5.XWalkViewTestAsync" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="12">
         <description>
           <test_script_entry>org.xwalk.embedding.test.v5.XWalkViewTestAsync</test_script_entry>
         </description>
