+++ embeddingapi/embedding-asyncapi-android-tests/tests_v6.xml
@@ -8,6 +8,11 @@
           <test_script_entry>org.xwalk.embedding.test.v6.XWalkSettingTestAsync</test_script_entry>
         </description>
       </testcase>
+      <testcase component="Crosswalk APIs/Embedding API" execution_type="auto" id="v6.XWalkViewTestAsync" purpose="Check if the methods of XWalkView interface can be executed correctly." subcase="3">
+        <description>
+          <test_script_entry>org.xwalk.embedding.test.v6.XWalkViewTestAsync</test_script_entry>
+        </description>
+      </testcase>
     </set>
   </suite>
 </test_definition>
